+++ test/watchers/NetworkWatcher.test.ts
@@ -3,8 +3,10 @@ import https from "https";
 import { NetworkWatcher, Watcher } from "../../src/watchers";
 import { Agent } from "../../src/Agent";
 import { AgentRegistrar } from "../../src/AgentRegistrar";
+import { TrackJSOptions } from "../../src/types";
+import { TrackJSEntry } from "../../src/types/TrackJSCapturePayload";
 
-let _NetworkWatcher = NetworkWatcher as Watcher;
+const _NetworkWatcher = NetworkWatcher as Watcher;
 
 beforeAll(() => {
   Agent.defaults.dependencies = false;
@@ -14,42 +16,93 @@ jest.mock("net");
 
 describe("NetworkWatcher", () => {
   describe("install()", () => {
+    let fakeOptions: TrackJSOptions;
     let fakeAgent: Agent;
 
     beforeEach(() => {
+      fakeOptions = { network: { error: true, enabled: true } };
       fakeAgent = new Agent({ token: "test" });
       AgentRegistrar.init(fakeAgent);
-      _NetworkWatcher.install();
     });
 
     afterEach(() => {
       _NetworkWatcher.uninstall();
     });
 
-    it("http patched to intercept request", async () => {
+    it("when disabled, does nothing", (done) => {
+      _NetworkWatcher.install({ network: { enabled: false } });
       fakeAgent.telemetry.add = jest.fn();
-      await http.request("http://example.com/?foo=bar");
-      expect(fakeAgent.telemetry.add).toHaveBeenCalledWith(
-        "n",
-        expect.objectContaining({
-          method: "GET",
-          url: "http://example.com/?foo=bar",
-          startedOn: expect.any(String)
-        })
-      );
+      http.get("http://example.com/?foo=bar", () => {
+        expect(fakeAgent.telemetry.add).not.toHaveBeenCalled();
+        done();
+      });
     });
 
-    it("https patched to intercept request", async () => {
+    it("when enabled, http patched to intercept request", (done) => {
+      _NetworkWatcher.install({ network: { enabled: true } });
       fakeAgent.telemetry.add = jest.fn();
-      await https.request("https://example.com/?foo=bar");
-      expect(fakeAgent.telemetry.add).toHaveBeenCalledWith(
-        "n",
-        expect.objectContaining({
-          method: "GET",
-          url: "https://example.com/?foo=bar",
-          startedOn: expect.any(String)
-        })
-      );
+
+      const req = http.request("http://example.com/?foo=bar", { method: "GET" }, (resp) => {
+        expect(fakeAgent.telemetry.add).toHaveBeenCalledWith(
+          "n",
+          expect.objectContaining({
+            method: "GET",
+            url: "http://example.com/?foo=bar",
+            startedOn: expect.any(String)
+          })
+        );
+        done();
+      });
+      req.end();
     });
+
+    it("when enabled, https patched to intercept request", (done) => {
+      _NetworkWatcher.install(fakeOptions);
+      fakeAgent.telemetry.add = jest.fn();
+      const req = https.request("https://example.com/?foo=bar", { method: "GET" }, (res) => {
+        expect(fakeAgent.telemetry.add).toHaveBeenCalledWith(
+          "n",
+          expect.objectContaining({
+            method: "GET",
+            url: "https://example.com/?foo=bar",
+            startedOn: expect.any(String)
+          })
+        );
+        done();
+      });
+      req.end();
+    });
+
+    it("when error enabled, captures error from failing request", (done) => {
+      _NetworkWatcher.install({ network: { enabled: true, error: true }});
+      fakeAgent.telemetry.add = jest.fn();
+      fakeAgent.captureError = jest.fn();
+      const req = https.request("https://httpstat.us/501", { method: "GET" }, (res) => {
+        setTimeout(() => {
+          expect(fakeAgent.captureError).toHaveBeenCalledWith(
+            expect.objectContaining({
+              stack: expect.any(String)
+            }),
+            TrackJSEntry.Network
+          );
+          done();
+        });
+      });
+      req.end();
+    });
+
+    it("when error disabled, no errors captured", (done) => {
+      _NetworkWatcher.install({ network: { enabled: true, error: false }});
+      fakeAgent.telemetry.add = jest.fn();
+      fakeAgent.captureError = jest.fn();
+      const req = https.request("https://httpstat.us/501", { method: "GET" }, (res) => {
+        setTimeout(() => {
+          expect(fakeAgent.captureError).not.toHaveBeenCalled();
+          done();
+        });
+      });
+      req.end();
+    });
+
   });
 });